requirements.txt
accelerate==0.30.1
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
dataclasses-json==0.6.4
datasets==2.14.6
dill==0.3.7
distro==1.9.0
dspy==0.1.5
dspy-ai==2.4.9
exceptiongroup==1.2.0
fastembed==0.2.7
filelock==3.13.1
flatbuffers==24.3.7
frozenlist==1.4.1
fsspec==2023.10.0
grpcio==1.62.1
grpcio-tools==1.62.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx==0.25.2
huggingface-cli==0.1
huggingface-hub==0.20.3
humanfriendly==10.0
hyperframe==6.0.1
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langchain-mistralai==0.0.5
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langsmith==0.1.31
loguru==0.7.2
Mako==1.3.2
MarkupSafe==2.1.5
marshmallow==3.21.1
mistralai==0.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
ollama==0.1.8
onnx==1.15.0
onnxruntime==1.17.1
openai==1.16.2
optuna==3.4.0
orjson==3.9.15
packaging==23.2
pandas==2.2.2
pillow==10.2.0
portalocker==2.8.2
protobuf==4.25.3
psutil==5.9.8
pyarrow==15.0.2
pyarrow-hotfix==0.6
pydantic==2.7.1
pydantic_core==2.18.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.9.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.7.0
sentencepiece==0.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
structlog==24.1.0
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.1
transformers==4.39.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
ujson==5.8.0
urllib3==1.26.18
xxhash==3.4.1
yarl==1.9.4
